二维码收款结果、重构账单

+++ AggregatePay/Business/Collection/QRCodeCollection/Controllers/QRCodePayElement/Controller/APQRCPElementViewController.swift
@@ -117,7 +117,7 @@ extension APQRCPElementViewController {
     }
     
     private func startHttpGetQRCode() {
-        qrCodePayRequest.transAmount = String((Int(amountStr!)! * 100))
+        qrCodePayRequest.transAmount = String((Double(amountStr!)! * 100))
         qrCodePayRequest.userId = APUserDefaultCache.AP_get(key: .userId) as! String
         view.AP_loadingBegin()
         APNetworking.post(httpUrl: APHttpUrl.trans_httpUrl,
